[GPXM-3] feat(geolocation): add async Nominatim client
